null checks to fix bugs
 - around source zip cleanup (if user configured no files for the src
   upload, the zip util would return null and fail the build when trying
   to delete it)
 - around a call to response.getJobId, since IntelliJ warned that
   response might be null at that point

+++ src/main/java/org/jenkinsci/plugins/codedx/CodeDxPublisher.java
@@ -337,7 +337,9 @@ public class CodeDxPublisher extends Recorder {
 					do {
 						Thread.sleep(3000);
 						oldStatus = status;
-						status = repeatingClient.getJobStatus(response.getJobId());
+						if(response != null) {
+							status = repeatingClient.getJobStatus(response.getJobId());
+						}
 						if (status != null && !status.equals(oldStatus)) {
 							if (Job.QUEUED.equals(status)) {
 								buildOutput.println("Code Dx analysis is queued");
@@ -438,7 +440,9 @@ public class CodeDxPublisher extends Recorder {
 			} catch (NumberFormatException e) {
 				buildOutput.println("Invalid project Id");
 			} finally {
-				sourceAndBinaryZip.delete();
+				if(sourceAndBinaryZip != null){
+					sourceAndBinaryZip.delete();
+				}
 			}
 		} else {
 			buildOutput.println("Nothing to send, this doesn't seem right! Please check your 'Code Dx > Source and Binary Files' configuration.");
